requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.21.1
gunicorn==20.1.0
SQLAlchemy==2.0.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
emails==0.6
psycopg2-binary==2.9.6
apscheduler==3.10.1
sqlalchemy_mptt==0.2.5
sqlalchemy_mixins @ git+https://github.com/aurthurm/sqlalchemy-mixins.git@async_support
tenacity==8.2.2
pydantic[email]==1.10.7
python-multipart==0.0.6
alembic==1.10.4
asyncpg==0.27.0
strawberry-graphql==0.174.0
broadcaster~=0.2.0
aiofiles==23.1.0
pandas==2.0.1
jinja2==3.1.2
celery==5.2.7
fpdf2==2.7.3
PyPDF2==3.0.1
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-instrumentation-sqlalchemy==0.38b0
